+++ index.html
@@ -149,9 +149,5 @@
   </section>
 </template>
 <script src="js/main.js"></script>
-<script src="js/player.js"></script>
-<script src="js/animate.js"></script>
-<script src="js/time-format.js"></script>
-<script src="js/timer.js"></script>
 </body>
 </html>
