+++ js/gameData.js
@@ -1,5 +1,5 @@
-export default Object.freeze({
+export default {
   genre: [`Хип-хоп и рэп`, `Классические`, `Кантри и фолк`, `Поп`],
   artist: [`Twin Musicom`, `Mozart`, `Unicorn Heads`, `Silent Partner`],
   url: [`https://www.youtube.com/audiolibrary_download?vid=1d01174c384ac88c`, `https://www.youtube.com/audiolibrary_download?vid=5f37e1a59d003588`, `https://www.youtube.com/audiolibrary_download?vid=4bae71dd4663c88d`, `https://www.youtube.com/audiolibrary_download?vid=47826af5576d3ae4`]
-});
+};
