+++ js/gameRules.js
@@ -1,4 +1,5 @@
 export default {
+  gamesNumber: 10,
   time: 2,
   artistsNumber: 3,
   genreNumber: 4
