+++ js/main.js
@@ -1,3 +1,7 @@
+import './animate.js';
+import './player.js';
+import './time-format.js';
+import './timer.js';
 import renderWelcome from './renderWelcome';
 
 renderWelcome();
