+++ js/renderNextScreen.js
@@ -1,3 +1,4 @@
+import gameRules from './gameRules';
 import currentState from './currentState';
 import renderGenre from './renderGenre';
 import renderArtist from './renderArtist';
@@ -5,7 +6,7 @@ import renderWin from './renderWin';
 
 export default () => {
   currentState.answerCount++;
-  if (currentState.answerCount === 11) {
+  if (currentState.answerCount === gameRules.gamesNumber + 1) {
     currentState.answerCount = 0;
     renderWin(renderArtist.rightAnswerCount);
   } else {
