+++ js/screenWin.js
@@ -1,0 +1,13 @@
+import elementCreator from './elementCreator';
+
+const winTemplate = (data) => `
+  <section class="main main--result">
+    <section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
+    <h2 class="title">Вы настоящий меломан!</h2>
+    <div class="main-stat">${data.rightAnswerCount}</div>
+    <span class="main-comparison">Это&nbsp;лучше чем у&nbsp;80%&nbsp;игроков</span>
+    <span role="button" tabindex="0" class="main-replay">Сыграть ещё раз</span>
+  </section>
+`;
+
+export default (data) => elementCreator(winTemplate(data));
